Remove `JournalAdminController` and `JournalItemReadRepository`, update `Domain` namespace, and refactor main repository logic for `JournalItem`.
