+++ src/Bitrix24Accounts/UseCase/ChangeDomainUrl/Command.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Bitrix24\Lib\Bitrix24Accounts\UseCase\ChangeDomainUrl;
 
-use Bitrix24\Lib\Kernel\ValueObjects\Domain;
+use Bitrix24\Lib\Common\ValueObjects\Domain;
 
 readonly class Command
 {
