+++ src/Journal/Infrastructure/Doctrine/DoctrineDbalJournalItemRepository.php
@@ -13,6 +13,9 @@ declare(strict_types=1);
 
 namespace Bitrix24\Lib\Journal\Infrastructure\Doctrine;
 
+use Bitrix24\Lib\ApplicationInstallations\Entity\ApplicationInstallation;
+use Bitrix24\Lib\Bitrix24Accounts\Entity\Bitrix24Account;
+use Bitrix24\Lib\Common\ValueObjects\Domain;
 use Bitrix24\Lib\Journal\Entity\JournalItem;
 use Bitrix24\Lib\Journal\Entity\JournalItemInterface;
 use Bitrix24\Lib\Journal\Entity\LogLevel;
@@ -20,6 +23,9 @@ use Bitrix24\Lib\Journal\Infrastructure\JournalItemRepositoryInterface;
 use Carbon\CarbonImmutable;
 use Doctrine\ORM\EntityManagerInterface;
 use Doctrine\ORM\EntityRepository;
+use Doctrine\ORM\QueryBuilder;
+use Knp\Component\Pager\Pagination\PaginationInterface;
+use Knp\Component\Pager\PaginatorInterface;
 use Symfony\Component\Uid\Uuid;
 
 class DoctrineDbalJournalItemRepository implements JournalItemRepositoryInterface
@@ -27,7 +33,8 @@ class DoctrineDbalJournalItemRepository implements JournalItemRepositoryInterfac
     private readonly EntityRepository $repository;
 
     public function __construct(
-        private readonly EntityManagerInterface $entityManager
+        private readonly EntityManagerInterface $entityManager,
+        private readonly PaginatorInterface $paginator
     ) {
         $this->repository = $this->entityManager->getRepository(JournalItem::class);
     }
@@ -132,5 +139,116 @@ class DoctrineDbalJournalItemRepository implements JournalItemRepositoryInterfac
             ->getQuery()
             ->execute()
         ;
+    }
+
+    /**
+     * Find journal items with filters and pagination.
+     *
+     * @return PaginationInterface<JournalItemInterface>
+     */
+    public function findWithFilters(
+        ?string $memberId = null,
+        ?Domain $domain = null,
+        ?LogLevel $logLevel = null,
+        ?string $label = null,
+        int $page = 1,
+        int $limit = 50
+    ): PaginationInterface {
+        $queryBuilder = $this->createFilteredQueryBuilder($memberId, $domain, $logLevel, $label);
+
+        return $this->paginator->paginate(
+            $queryBuilder,
+            $page,
+            $limit,
+            [
+                'defaultSortFieldName' => 'j.createdAt',
+                'defaultSortDirection' => 'desc',
+            ]
+        );
+    }
+
+    /**
+     * Get available domain URLs from journal.
+     *
+     * @return string[]
+     */
+    public function getAvailableDomains(): array
+    {
+        // Join with ApplicationInstallation and then Bitrix24Account to get domain URLs
+        $queryBuilder = $this->entityManager->createQueryBuilder();
+        $queryBuilder->select('DISTINCT b24.domainUrl')
+            ->from(JournalItem::class, 'j')
+            ->innerJoin(ApplicationInstallation::class, 'ai', 'WITH', 'ai.id = j.applicationInstallationId')
+            ->innerJoin(Bitrix24Account::class, 'b24', 'WITH', 'b24.id = ai.bitrix24AccountId')
+            ->orderBy('b24.domainUrl', 'ASC')
+        ;
+
+        $results = $queryBuilder->getQuery()->getScalarResult();
+
+        return array_column($results, 'domainUrl');
+    }
+
+    /**
+     * Get available labels from journal.
+     *
+     * @return string[]
+     */
+    public function getAvailableLabels(): array
+    {
+        $queryBuilder = $this->entityManager->createQueryBuilder();
+        $queryBuilder->select('DISTINCT j.label')
+            ->from(JournalItem::class, 'j')
+            ->where('j.label IS NOT NULL')
+            ->orderBy('j.label', 'ASC')
+        ;
+
+        $results = $queryBuilder->getQuery()->getScalarResult();
+
+        return array_filter(array_column($results, 'label'));
+    }
+
+    /**
+     * Create query builder with filters.
+     */
+    private function createFilteredQueryBuilder(
+        ?string $memberId = null,
+        ?Domain $domain = null,
+        ?LogLevel $logLevel = null,
+        ?string $label = null
+    ): QueryBuilder {
+        $queryBuilder = $this->entityManager->createQueryBuilder();
+        $queryBuilder->select('j')
+            ->from(JournalItem::class, 'j')
+        ;
+
+        if (null !== $memberId) {
+            $queryBuilder->andWhere('j.memberId = :memberId')
+                ->setParameter('memberId', $memberId)
+            ;
+        }
+
+        if (null !== $domain) {
+            $queryBuilder->innerJoin(ApplicationInstallation::class, 'ai', 'WITH', 'ai.id = j.applicationInstallationId')
+                ->innerJoin(Bitrix24Account::class, 'b24', 'WITH', 'b24.id = ai.bitrix24AccountId')
+                ->andWhere('b24.domainUrl = :domainUrl')
+                ->setParameter('domainUrl', $domain->value)
+            ;
+        }
+
+        if (null !== $logLevel) {
+            $queryBuilder->andWhere('j.level = :level')
+                ->setParameter('level', $logLevel)
+            ;
+        }
+
+        if (null !== $label) {
+            $queryBuilder->andWhere('j.label = :label')
+                ->setParameter('label', $label)
+            ;
+        }
+
+        $queryBuilder->orderBy('j.createdAt', 'DESC');
+
+        return $queryBuilder;
     }
 }
