+++ tests/Functional/Bitrix24Accounts/UseCase/InstallStart/HandlerTest.php
@@ -33,7 +33,7 @@ use Random\RandomException;
 use Symfony\Component\EventDispatcher\Debug\TraceableEventDispatcher;
 use Symfony\Component\EventDispatcher\EventDispatcher;
 use Symfony\Component\Stopwatch\Stopwatch;
-use Bitrix24\Lib\Kernel\ValueObjects\Domain;
+use Bitrix24\Lib\Common\ValueObjects\Domain;
 
 /**
  * @internal
