+++ tests/Unit/ApplicationInstallations/UseCase/OnAppInstall/CommandTest.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace Bitrix24\Lib\Tests\Unit\ApplicationInstallations\UseCase\OnAppInstall;
 
 use Bitrix24\Lib\ApplicationInstallations\UseCase\OnAppInstall\Command;
-use Bitrix24\Lib\Kernel\ValueObjects\Domain;
+use Bitrix24\Lib\Common\ValueObjects\Domain;
 use Bitrix24\Lib\Tests\Functional\ApplicationInstallations\Builders\ApplicationInstallationBuilder;
 use Bitrix24\Lib\Tests\Functional\Bitrix24Accounts\Builders\Bitrix24AccountBuilder;
 use Bitrix24\SDK\Application\ApplicationStatus;
